Making test_properties into its own util: BCSS-20024 (#15)
<!-- markdownlint-disable-next-line first-line-heading -->

<!-- Describe your changes in detail. -->
This is moving all of the smokescreen_properties/test_properties methods
into a util to be used by all tests.
As there are two different property files, one for the compartments and
one for the rest I added an if statement to select between them
depending on the input provided

This change is tracked by the Jira ticket:
[BCSS-20024](https://nhsd-jira.digital.nhs.uk/browse/BCSS-20024)

<!-- Why is this change required? What problem does it solve? -->
This helps to reduce the number of duplicate code and we can easier load
the properties file across all of our tests with the same method,
instead of repeating it again for each test.

<!-- What types of changes does your code introduce? Put an `x` in all
the boxes that apply. -->

- [x] Refactoring (non-breaking change)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would change existing
functionality)
- [ ] Bug fix (non-breaking change which fixes an issue)

<!-- Go over all the following points, and put an `x` in all the boxes
that apply. -->

- [x] I am familiar with the [contributing
guidelines](https://github.com/nhs-england-tools/playwright-python-blueprint/blob/main/CONTRIBUTING.md)
- [x] I have followed the code style of the project
- [ ] I have added tests to cover my changes (where appropriate)
- [x] I have updated the documentation accordingly
- [ ] This PR is a result of pair or mob programming

---

To ensure the utmost confidentiality and protect your and others
privacy, we kindly ask you to NOT including [PII (Personal Identifiable
Information) / PID (Personal Identifiable
Data)](https://digital.nhs.uk/data-and-information/keeping-data-safe-and-benefitting-the-public)
or any other sensitive data in this PR (Pull Request) and the codebase
changes. We will remove any PR that do contain any sensitive
information. We really appreciate your cooperation in this matter.

- [x] I confirm that neither PII/PID nor sensitive data are included in
this PR and the codebase changes.