Adding scenarios 15 & 16 from feature file (#119)
<!-- markdownlint-disable-next-line first-line-heading -->
## Description

<!-- Describe your changes in detail. -->
Migrating scenario 15 - 16 from the
EndoscopyInvestigationDatasetScenarios feature file:

15: 
Scenario: Check the validation of the Scope ID field in the Endoscopy
Information section in an incomplete Colonoscopy investigation dataset
16:
Scenario: Check the behaviour of the Endoscope Inserted fields in the
Endoscopy Information section in an incomplete Colonoscopy investigation
dataset

## Context

<!-- Why is this change required? What problem does it solve? -->
Migrating scenario 15 - 16 from the
EndoscopyInvestigationDatasetScenarios feature file:

15: 
Scenario: Check the validation of the Scope ID field in the Endoscopy
Information section in an incomplete Colonoscopy investigation dataset
16:
Scenario: Check the behaviour of the Endoscope Inserted fields in the
Endoscopy Information section in an incomplete Colonoscopy investigation
dataset

## Type of changes

<!-- What types of changes does your code introduce? Put an `x` in all
the boxes that apply. -->

- [x] Refactoring (non-breaking change)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would change existing
functionality)
- [ ] Bug fix (non-breaking change which fixes an issue)

## Checklist

<!-- Go over all the following points, and put an `x` in all the boxes
that apply. -->

- [x] I am familiar with the [contributing
guidelines](https://github.com/nhs-england-tools/playwright-python-blueprint/blob/main/CONTRIBUTING.md)
- [x] I have followed the code style of the project
- [x] I have added tests to cover my changes (where appropriate)
- [x] I have updated the documentation accordingly
- [ ] This PR is a result of pair or mob programming

---

## Sensitive Information Declaration

To ensure the utmost confidentiality and protect your and others
privacy, we kindly ask you to NOT including [PII (Personal Identifiable
Information) / PID (Personal Identifiable
Data)](https://digital.nhs.uk/data-and-information/keeping-data-safe-and-benefitting-the-public)
or any other sensitive data in this PR (Pull Request) and the codebase
changes. We will remove any PR that do contain any sensitive
information. We really appreciate your cooperation in this matter.

- [x] I confirm that neither PII/PID nor sensitive data are included in
this PR and the codebase changes.